separate into `speed.go` file
